modified templates && little refactoring
