+++ src/main/java/com/WWI16AMA/backend_api/Email/EmailService.java
@@ -3,9 +3,7 @@ package com.WWI16AMA.backend_api.Email;
 import com.WWI16AMA.backend_api.Account.Transaction;
 import com.WWI16AMA.backend_api.Member.Member;
 import org.apache.commons.io.IOUtils;
-import org.apache.tomcat.jni.Local;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.InputStreamSource;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -19,14 +17,10 @@ import org.thymeleaf.context.Context;
 
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
-
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.net.URL;
-import java.time.LocalDate;
-import java.util.Date;
 import java.util.Locale;
 
 @Service
@@ -40,30 +34,20 @@ public class EmailService {
 
 
     public void sendBillingNotification(
-            final Member member, Locale locale, String subjectBody, String contact, Transaction transaction)
+            final Member member, Locale locale, Transaction transaction)
             throws MessagingException {
 
         final MultipartFile multipartFile = getLogo();
-        Context billingCtx = prepareBasicData(locale);
-        billingCtx.setVariable("subjectBody", subjectBody);
-        billingCtx.setVariable("contact", contact);
-        billingCtx.setVariable("name", member.getFirstName()+" " +member.getLastName());
-        billingCtx.setVariable("amount", -transaction.getAmount());
-        billingCtx.setVariable("bankingAccount", member.getBankingAccount());
-        billingCtx.setVariable("memberID", member.getId());
-        billingCtx.setVariable("date", LocalDate.now());
-        billingCtx.setVariable("bankName", "leer");
-        billingCtx.setVariable("bic", "leer");
-        billingCtx.setVariable("jobNumber", "leer");
 
-        billingCtx.setVariable("imageResourceName", multipartFile.getName());
-
+        Context billingCtx = prepareBasicData(locale, member, multipartFile);
+        billingCtx.setVariable("amount", formatAmount(transaction.getAmount()));
 
         final MimeMessage mimeMessage = this.javaMailSender.createMimeMessage();
         final MimeMessageHelper message =
                 new MimeMessageHelper(mimeMessage, true, "UTF-8"); // true = multipart
-        message.setSubject("Aufwandsentschädigung");
+        message.setSubject("Abbuchung Ihres Mitgliedsbeitrages");
         message.setFrom("Flugverein Reilingen <flugverein@reilingen.com>");
+        //message.setTo(member.getEmail());
         message.setTo("him.kourouma@hotmail.com");
 
         final String htmlContent = this.htmlTemplateEngine.process("billing-email.html", billingCtx);
@@ -80,12 +64,9 @@ public class EmailService {
 
         this.javaMailSender.send(mimeMessage);
 
-
     }
 
-    public MultipartFile getLogo()  {
-
-
+    public MultipartFile getLogo() {
         File file = null;
         FileInputStream input = null;
         MultipartFile multipartFile = null;
@@ -104,10 +85,20 @@ public class EmailService {
         return multipartFile;
     }
 
-    private Context prepareBasicData(Locale locale){
+    private Context prepareBasicData(Locale locale, Member member, MultipartFile multipartFile) {
         final Context ctx = new Context(locale);
+        ctx.setVariable("lastName", member.getLastName());
+        ctx.setVariable("imageResourceName", multipartFile.getName());
 
         return ctx;
+    }
+
+    private String formatAmount(double amount) {
+        if (amount < 0) {
+            return (-amount) + "0"; // Negatives Vorzeichen wird für die E-Mail entfernt.
+        } else {
+            return amount + "0";
+        }
     }
 
 }
