+++ src/main/resources/templates/betankung-email.html
@@ -8,7 +8,6 @@
 </head>
 <body>
 <!--[if (gte mso 9)|(IE)]>
-<![endif]-->
 <table class="size" align="center" cellpadding="0" cellspacing="0" border="0" bgcolor="white">
     <tr>
         <td class="header">
@@ -23,15 +22,14 @@
     </tr>
     <tr>
         <td class="content">
-            <br/>
-            wir freuen uns, dass Sie auch dieses Jahr Mitglied im Flugsportverein Reilingen sind.<br/>
-            Ihr Mitgliedsbeitrag in Höhe von <span th:text="${amount}"/>€ wurde von ihrem Mitgliedskonto abgebucht.
+            vielen Dank für das Betanken des Flugzeuges <span th:text="${planeName}"/>.<br/>
+            Der Betrag in Höhe von <span th:text="${amount}"/>€ wurde ihrem Mitgliedskonto gutgeschrieben
             <br/>
             <br/>
             Freundliche Grüße,
             <br/>
-            der Finanzverwalter des Flugsportvereins Reilingen e.V.
             <br/>
+            der Finanzverwalter des Flugsportvereins Reilingen e.V.
             <br/>
         </td>
     </tr>
@@ -39,7 +37,8 @@
     <br/>
     <tr>
         <td class="footer" bgcolor="#0081C9">
-            Dies ist eine automatisch generierte E-Mail. Bitte antworten Sie nicht drauf, da uns Ihre Nachricht auf diesem Weg nicht erreichen wird.
+            Dies ist eine automatisch generierte E-Mail. Bitte antworten Sie nicht drauf, da uns Ihre Nachricht auf diesem Weg nicht erreichen wird. Sollten Sie Fragen zu dem Inhalt dieser E-Mail haben, wenden Sie sich bitte driekt an den in der E-Mail erwähnten Ansprechpartner.
+        </td>
     </tr>
 </table>
 <!--[if (gte mso 9)|(IE)]>
